Change all bytes to int. Add InctidentType in Incident table. Add IncifentType Enum to Dictionaries.cs .

+++ SRC/Application/Players/Commands/UpdatePlayer/UpdatePlayerCommand.cs
@@ -16,8 +16,8 @@ namespace BasketballLeague.Application.Players.Commands.UpdatePlayer
         public string Surname { get; set; }
         public DateTime? Birthdate { get; set; }
         public string PhotoUrl { get; set; }
-        public byte? Height { get; set; }
-        public byte? Position { get; set; }
+        public int? Height { get; set; }
+        public int? Position { get; set; }
 
         public class Handler : IRequestHandler<UpdatePlayerCommand>
         {
