+++ SRC/Domain/Entities/Player.cs
@@ -10,8 +10,8 @@ namespace BasketballLeague.Domain.Entities
         public string Surname { get; set; }
         public DateTime? Birthdate { get; set; }
         public string PhotoUrl { get; set; }
-        public byte? Height { get; set; }
-        public byte? Position { get; set; }
+        public int? Height { get; set; }
+        public int? Position { get; set; }
 
         public virtual ICollection<Assist> Assists { get; set; }
         public virtual ICollection<Block> Blocks { get; set; }
